+++ Axis.Luna.Common.Test/SecureRandomTests.cs
@@ -3,6 +3,7 @@ using Axis.Luna.Common.Utils;
 using Axis.Luna.Extensions;
 using Microsoft.VisualStudio.TestTools.UnitTesting;
 using System;
+using System.Linq;
 using System.Text;
 
 namespace Axis.Luna.Common.Test